+++ src/rx-document.js
@@ -379,10 +379,10 @@ export const basePrototype = {
      */
     remove() {
         if (this.deleted) {
-            throw RxError.newRxError('DOC13', {
+            return Promise.reject(RxError.newRxError('DOC13', {
                 document: this,
                 id: this.primary
-            });
+            }));
         }
 
         const deletedData = clone(this._data);